Refactoring with multiple files for each slice.
