+++ src/components/Auth.js
@@ -1,5 +1,5 @@
 import { useDispatch } from 'react-redux';
-import { authActions } from '../store/index';
+import { authActions } from '../store/auth-slice';
 
 import classes from './Auth.module.css';
