+++ src/components/Counter.js
@@ -2,7 +2,7 @@
 // import { connect } from 'react-redux';
 import { useSelector, useDispatch } from 'react-redux';
 
-import { counterActions } from '../store';
+import { counterActions } from '../store/counter-slice';
 import classes from './Counter.module.css';
 
 const Counter = () => {
