+++ src/components/Header.js
@@ -1,6 +1,6 @@
 import { useSelector, useDispatch } from 'react-redux';
 
-import { authActions } from '../store/index';
+import { authActions } from '../store/auth-slice';
 
 import classes from './Header.module.css';
