+++ src/store/index.js
@@ -1,41 +1,7 @@
 // import { createStore } from 'redux';
-import { createSlice, configureStore } from '@reduxjs/toolkit';
-
-const initialCounterState = { counter: 0, showCounter: true };
-
-const counterSlice = createSlice({
-  name: 'counter',
-  initialState: initialCounterState,
-  reducers: {
-    increment(state) {
-      state.counter++; // Here you can mutate the state, although Redux make it immutable behind the scenes
-    },
-    decrement(state) {
-      state.counter--;
-    },
-    increaseBy5(state, action) {
-      state.counter += action.payload;
-    },
-    toggleCounter(state) {
-      state.showCounter = !state.showCounter
-    }
-  }
-})
-
-const initialAuthState = { isAuthenticated: false };
-
-const authSlice = createSlice({
-  name: 'auth',
-  initialState: initialAuthState,
-  reducers: {
-    login(state) {
-      state.isAuthenticated = true
-    },
-    logout(state) {
-      state.isAuthenticated = false
-    }
-  }
-})
+import { configureStore } from '@reduxjs/toolkit';
+import counterReducer from './counter-slice';
+import authReducer from './auth-slice';
 
 // const counterReducer = (state = initialState, action) => {
 //   switch (action.type) {
@@ -67,10 +33,7 @@ const authSlice = createSlice({
 // const store = createStore(counterReducer);
 
 const store = configureStore({
-  reducer: { counter: counterSlice.reducer, auth: authSlice.reducer },
+  reducer: { counter: counterReducer, auth: authReducer },
 });
-
-export const counterActions = counterSlice.actions;
-export const authActions = authSlice.actions;
 
 export default store;
